+++ packages/cw-orch-mock/src/core.rs
@@ -130,6 +130,18 @@ impl Mock<MockState> {
     pub fn new(sender: &Addr) -> Self {
         Mock::new_custom(sender, MockState::new())
     }
+
+    pub fn with_chain_id(sender: &Addr, chain_id: Option<&str>) -> Self {
+        let chain = Mock::new_custom(sender, MockState::new());
+        if let Some(chain_id) = chain_id {
+            chain
+                .app
+                .borrow_mut()
+                .update_block(|b| b.chain_id = chain_id.to_string());
+        }
+
+        chain
+    }
 }
 
 impl<S: StateInterface> Mock<S> {