+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## Unpublished
 
 - Ability to use the ExecuteFns and QueryFns traits on Units and Unnamed enum variants by @Kayanski
+- Added Sei Network
 - Added optional disable of sorting of ExecuteFns and QueryFns fields by @Kayanski
-
 
 ## v0.15.0