+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Ability to use the ExecuteFns and QueryFns traits on Units and Unnamed enum variants by @Kayanski
 - Added Sei Network
+- Make the broadcast function public on Sender
 
 ## v0.15.0