+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unpublished
 
 - Ability to use the ExecuteFns and QueryFns traits on Units and Unnamed enum variants by @Kayanski
+- Added Sei Network
+- Make the broadcast function public on Sender
+- Added a `call_as` helper on TxHandler Trait
+- Added optional disable of sorting of ExecuteFns and QueryFns fields by @Kayanski
 
 
 ## v0.15.0