+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Added helper to modify the chain id of cw-multi-test (Mock)
 - Added a trait to be able to commit any transaction to chain (Protobuf any type)
 - Added min gas and average gas utilization for computing the tx fee
+- Added Install Readme 
 
 ## v0.15.0